Move Logfire shutdown logic to config
