+++ logfire/_internal/main.py
@@ -10,7 +10,6 @@ from contextlib import AbstractContextManager
 from contextvars import Token
 from enum import Enum
 from functools import cached_property, partial
-from time import time
 from typing import (  # NOQA UP035
     TYPE_CHECKING,
     Any,
@@ -2480,52 +2479,7 @@ class Logfire:
         Returns:
             `False` if the timeout was reached before the shutdown was completed, `True` otherwise.
         """
-        start = time()
-
-        def remaining_ms() -> int:
-            return max(0, int(timeout_millis - (time() - start) * 1000))
-
-        complete = True
-        self.config.get_variable_provider().shutdown(timeout_millis=remaining_ms())
-        remaining = remaining_ms()
-        if not remaining:  # pragma: no cover
-            complete = False
-
-        forwarding_shutdown_result = self.config._otlp_forwarding.shutdown(  # pyright: ignore[reportPrivateUsage]
-            remaining, drain_queued=flush
-        )
-        complete = complete and forwarding_shutdown_result
-        remaining = remaining_ms()
-        if not remaining:  # pragma: no cover
-            complete = False
-
-        if flush and remaining:  # pragma: no branch
-            self._tracer_provider.force_flush(remaining)
-            remaining = remaining_ms()
-            if not remaining:  # pragma: no cover
-                complete = False
-        elif flush:  # pragma: no cover
-            complete = False
-
-        self._tracer_provider.shutdown()
-        remaining = remaining_ms()
-        if not remaining:  # pragma: no cover
-            complete = False
-
-        if flush and remaining:  # pragma: no branch
-            self._meter_provider.force_flush(remaining)
-            remaining = remaining_ms()
-            if not remaining:  # pragma: no cover
-                complete = False
-        elif flush:  # pragma: no cover
-            complete = False
-
-        self._meter_provider.shutdown(remaining)
-        remaining = remaining_ms()
-        if not remaining:  # pragma: no cover
-            complete = False
-
-        return remaining_ms() > 0 and complete
+        return self._config.shutdown(timeout_millis=timeout_millis, flush=flush)
 
     @overload
     def var(
