+++ specs/browser-telemetry-forwarding-first-pass/code-spec.md
@@ -24,8 +24,8 @@ class LogfireConfig(_LogfireConfigData):
 
 `_otlp_forwarding` is the private owner for forwarding queues and transport resources associated with this config. Public forwarding helpers reach it through the selected `Logfire` instance's config; normal instrumentation exporters do not call it.
 
-**`logfire/_internal/main.py` shuts down forwarding with the Logfire instance.** *(implements "Forwarding participates in Logfire flush and shutdown", "Post-shutdown forwarding calls are locally dropped")*
-`Logfire.shutdown()` includes forwarding manager shutdown in the same overall deadline accounting used for variables, traces, and metrics for queued memory drain. With `flush=True`, shutdown closes forwarding admission before draining memory work, then waits for any active immediate forwarding send to finish before closing forwarding sessions and returning. With `flush=False`, shutdown still closes forwarding admission and sessions, but it passes `drain_queued=False` so queued forwarding work is dropped rather than flushed.
+**`LogfireConfig.shutdown()` owns forwarding shutdown with the config lifecycle.** *(implements "Forwarding participates in Logfire flush and shutdown", "Post-shutdown forwarding calls are locally dropped")*
+`Logfire.shutdown()` delegates to `LogfireConfig.shutdown()`, which includes forwarding manager shutdown in the same overall deadline accounting used for variables, traces, and metrics for queued memory drain. With `flush=True`, shutdown closes forwarding admission before draining memory work, then waits for any active immediate forwarding send to finish before closing forwarding sessions and returning. With `flush=False`, shutdown still closes forwarding admission and sessions, but it passes `drain_queued=False` so queued forwarding work is dropped rather than flushed.
 
 **`logfire/_internal/config.py` registers forwarding cleanup at exit and after fork.** *(implements "The forwarding worker is daemon and lifecycle-managed", "Forwarding reinitializes worker state after fork")*
 Configuration registers an atexit cleanup hook for forwarding managers after OpenTelemetry providers have registered their own shutdown hooks. The forwarding hook uses the default 30 second shutdown deadline shared across live config-owned forwarding managers. Forwarding managers also register an after-fork reinitializer following the OpenTelemetry batch processor pattern; the child process clears inherited forwarding queues and stale worker references so parent-process telemetry is not sent by the child.
@@ -341,8 +341,8 @@ TODO: Address the pre-existing `LogfireConfig.force_flush()` result aggregation 
 **`LogfireConfig.configure()` replaces forwarding lifecycle with configuration lifecycle.** *(implements "Forwarding transport lifecycle is owned by Logfire configuration, not by each forwarding call")*
 Reconfiguration collects forwarding destinations while the existing `send_to_logfire` branch resolves tokens and backend URLs, including credentials-derived values, from the same loop that constructs normal exporters. After exporter construction succeeds, configuration installs a fresh manager built from that destination list and closes the previous config-owned forwarding manager.
 
-**`Logfire.shutdown()` closes forwarding admission before provider shutdown completes.** *(implements "Forwarding participates in Logfire flush and shutdown", "Post-shutdown forwarding calls are locally dropped")*
-Shutdown calls the config-owned manager with the remaining deadline. With `flush=True`, the manager waits for forwarding queues and workers within that deadline and drops queued memory work if the deadline expires. With `flush=False`, shutdown still calls the manager, but passes `drain_queued=False` so queued memory work is dropped before waiting for workers. After this point, otherwise valid forwarding calls map to partial success and do not recreate pipelines.
+**`LogfireConfig.shutdown()` closes forwarding admission before provider shutdown completes.** *(implements "Forwarding participates in Logfire flush and shutdown", "Post-shutdown forwarding calls are locally dropped")*
+Shutdown calls the config-owned manager with the remaining deadline. With `flush=True`, the manager waits for forwarding queues and workers within that deadline and drops queued memory work if the deadline expires. With `flush=False`, shutdown still calls the manager, but passes `drain_queued=False` so queued memory work is dropped before waiting for workers. `Logfire.shutdown()` delegates to this config method. After this point, otherwise valid forwarding calls map to partial success and do not recreate pipelines.
 
 **Part 5: Response Representations**
