+++ tests/test_configure.py
@@ -652,17 +652,14 @@ def test_shutdown_otlp_forwarding_closes_local_forwarding_managers(
 
 
 def test_logfire_shutdown_closes_forwarding_without_drain_when_flush_false() -> None:
-    default_logfire = logfire.DEFAULT_LOGFIRE_INSTANCE
-    default_logfire.__dict__.pop('_tracer_provider', None)
-    default_logfire.__dict__.pop('_meter_provider', None)
-    config = default_logfire.config
+    config = logfire.DEFAULT_LOGFIRE_INSTANCE.config
     config._variable_provider = mock.Mock()  # pyright: ignore[reportPrivateUsage]
     config._otlp_forwarding = mock.Mock()  # pyright: ignore[reportPrivateUsage]
     config._otlp_forwarding.shutdown.return_value = True  # pyright: ignore[reportPrivateUsage]
     config._tracer_provider = mock.Mock()  # pyright: ignore[reportPrivateUsage]
     config._meter_provider = mock.Mock()  # pyright: ignore[reportPrivateUsage]
 
-    assert logfire.shutdown(timeout_millis=1000, flush=False) is True
+    assert config.shutdown(timeout_millis=1000, flush=False) is True
 
     forwarding_timeout = config._otlp_forwarding.shutdown.call_args.args[0]  # pyright: ignore[reportPrivateUsage]
     config._otlp_forwarding.shutdown.assert_called_once_with(forwarding_timeout, drain_queued=False)  # pyright: ignore[reportPrivateUsage]
